requirements.txt
altair==5.1.1
ansi2html==1.8.0
anyio==4.0.0
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backgroundremover==0.2.5
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
branca==0.6.0
Brlapi==0.8.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
cmake==3.27.5
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
commandlines==0.4.1
contourpy==1.1.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.2
dbus-python==1.2.18
debugpy==1.8.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.1.3
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.18.0
ffmpeg-python==0.2.0
filelock==3.12.4
filetype==1.2.0
Flask==2.2.5
folium==0.14.0
fonttools==4.42.1
fqdn==1.5.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.36
haversine==2.8.0
hsh==1.1.0
httplib2==0.20.2
idna==3.3
imageio==2.31.3
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lit==16.0.6
llvmlite==0.40.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.0.76
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
pil-plus==0.2
Pillow==9.0.1
platformdirs==3.10.0
plotly==5.17.0
plotly-express==0.4.1
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydeck==0.8.0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMatting==1.1.8
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
reportlab==3.6.8
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rpds-py==0.10.3
scikit-image==0.21.0
scipy==1.11.2
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.27.2
streamlit-elements==0.1.0
streamlit-folium==0.13.0
sympy==1.12
systemd-python==234
tenacity==8.2.3
terminado==0.17.1
tifffile==2023.9.18
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==2.0.5
usb-creator==0.3.7
validators==0.22.0
wadllib==1.3.6
waitress==2.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.3
widgetsnbextension==4.0.9
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.0
